requirements.txt
# requirements.txt
PyQt6>=6.4.0
openai>=1.0.0 
Markdown>=3.4.0
python-dotenv>=0.20.0
Pillow>=9.0.0
PyMuPDF>=1.20.0